devices: Manually update on/off status when action is performed.
This change will manually update the device state when we perform a turn
on/turn off action on one of the supported devices. Ideally, we would
need to do the same for all devices and supported actions, but I'm not
sure how that should be possible for devices like a climate or lock.