tuyapi: Use Session object for every request made. (#38)
* tuyapi: Use Session object for every request made.

This change implements the Session object, allowing us to reuse TCP
connections via requests pool.

* devices: Manually update on/off status when action is performed.

This change will manually update the device state when we perform a turn
on/turn off action on one of the supported devices. Ideally, we would
need to do the same for all devices and supported actions, but I'm not
sure how that should be possible for devices like a climate or lock.

* tuyapi: Implement request session as a class property.